Refactor play_game with method play
    - Instead of doing all the logic in play_game, it now only
loops and call play for each move.

    - Change the behaviour of collision checking:
the pacman moves, we check for collision, the ghosts move, we
check for collision. I think the gain in readability is worth
the possible loss in performance

    - Rewards are not an attribute of pacman anymore but rather
a local variable that is returned by each functions.
It is meant to mimic the classical RL environments where
reward = act(action)